refactor(tensilelite): narrow the fused-A2A flag slot to u32
The release flag was a u64 on an 8-byte stride because the SDMA ATOMIC
that raises it used ADD_RTN_64 (47). The stated reason -- "MORI ships
ADD64 but no ADD32" -- was wrong: it inferred a hardware capability from
MORI's vendored packet header, which defines only the one op MORI itself
uses. ATOMIC_OP is a 7-bit selector into the TC atomic op table, the same
op space the shader's global_atomic_* instructions use, and that table
carries a _RTN_32 and a _RTN_64 form of every op -- ADD_RTN_32 = 15,
ADD_RTN_64 = 47. 15 is exercised by SDMA's own MI350 verification suite.

The flag accumulates at most tokenTiles and the DRAIN poll already read
only the low dword, so 32 bits is ample. Switching the op drops the slot,
the host allocation (flagBytes) and the poll's lane offset from 8 bytes
to 4 together.

The non-returning ADD_32 = 0x4f was also tried on hardware: it is serviced
and passes the same 4-card run at an indistinguishable latency. 15 is kept
because it is the value the verification suite covers and 0x4f offered
nothing measurable in exchange -- this atomic fires once per (peer,
token-tile) pair, off the hot path.

Validated: pure-GEMM assembly byte-identical, so the non-fused path is
untouched; every line of the fused assembly diff attributable to this
change; 4-card numeric PASS on MED with drain=1 (recv 0 mismatches on all
four cards, race 10/10). The client had to be rebuilt for that run -- the
prebuilt one still over-allocates the flag buffer and would have masked
the only risk this change carries.

+++ projects/hipblaslt/tensilelite/client/src/FusedA2AClient.cpp
@@ -347,11 +347,10 @@ namespace TensileLite
             // constraint (AM/W)%MT0==0), so the contiguous feature extent is n_shard.
             const size_t nTokenPad = ((N + FUSED_A2A_N_TILE - 1) / FUSED_A2A_N_TILE) * FUSED_A2A_N_TILE;
             const size_t recvBytes    = (size_t)W * nTokenPad * nShard * sizeof(uint16_t); // bf16
-            // flag slots are u64, NOT u32: the release signal is an SDMA ATOMIC
-            // ADD64 (MORI's SDMA packet set has ADD64 and no ADD32), so each slot
-            // is written 8 bytes wide. With a 4-byte stride the top rank's atomic
-            // would run past the end of this allocation.
-            const size_t flagBytes    = (size_t)W * sizeof(uint64_t);
+            // One u32 flag slot per source rank, raised by an SDMA ATOMIC
+            // ADD_RTN_32. Must stay in step with emitComputeFlagAddr's *4 stride
+            // and the DRAIN poll's j*4.
+            const size_t flagBytes    = (size_t)W * sizeof(uint32_t);
             // counter is indexed [dst_rank][token-tile] -> W*tokenTiles u32 slots,
             // followed by a W-entry second-level counter2[dst_rank] (target
             // tokenTiles) at word index W*tokenTiles, then a single third-level
